support for layered config using a serde Deserializer as a source
This allows a serde Deserializer, representing a config file, to be used
as a value source, via the builder API, which is also new in this commit.

To use the serde feature, you have to put the `#[conf(serde)]` tag on your struct.

We also introduce a `ConfSerde` extension trait, which is implemented
when `#[conf(serde)]` is annotated, and provides implementation details
for this layered config feature on that struct.

The priority order is args > env > serde > defaults

We also support this in subcommands, via the `SubcommandsSerde` extension trait.

To further adjust the serde system, most fields have serde options like
`serde(rename)` and `serde(skip)`. Structs have a `serde(allow_unknown_fields)`
option.

Parameter and Repeat fields have a `serde(use_value_parser)` option, which
can be useful if you just want serde to read a string and then forward it
to the value parser.